update: move test select course to upload (#123)
