+++ Views/Upload/index.tsx
@@ -1,28 +1,29 @@
 import React, { useRef, useState } from "react";
-import { useRouter } from 'next/router'
+import { useRouter } from "next/router";
 import Modal from "../Modal";
 import { FaCloudUploadAlt } from "react-icons/fa";
 import { useAppContext } from "@/Context";
+import ModalSelectCourse from "@/components/common/ModalSelectCourse";
 
 export default function Upload() {
     const { setSrcVideoEdit, setThumbVideoEdit } = useAppContext();
     const { push } = useRouter();
     const [isModal, setIsModal] = useState<boolean>(false);
+    const [isModalSelectCourse, setModalSelectCourse] = useState<boolean>(true);
     const [isDragging, setIsDragging] = useState(false);
 
-    const canvasRef = useRef<HTMLCanvasElement>(null)
-    const videoRef = useRef<HTMLVideoElement>(null)
+    const canvasRef = useRef<HTMLCanvasElement>(null);
+    const videoRef = useRef<HTMLVideoElement>(null);
     const inputRef = useRef<HTMLInputElement>(null);
 
     function handleFileSelect(videoFile: File): boolean {
         const file = videoFile;
         const fileName = file.name;
-        const fileExtension: string = fileName.split('.').pop()!.toLocaleLowerCase();
-        const allowedExtensions: string[] = ['mp4', 'avi', 'mov'];
+        const fileExtension: string = fileName.split(".").pop()!.toLocaleLowerCase();
+        const allowedExtensions: string[] = ["mp4", "avi", "mov"];
         if (allowedExtensions.includes(fileExtension)) {
             return true;
-        }
-        else {
+        } else {
             return false;
         }
     }
@@ -32,15 +33,15 @@ export default function Upload() {
             let ctx = canvasRef.current.getContext("2d");
             if (ctx && videoRef.current) {
                 ctx.drawImage(videoRef.current, 0, 0, videoRef.current.videoWidth, videoRef.current.videoHeight);
-                setThumbVideoEdit(canvasRef.current.toDataURL())
+                setThumbVideoEdit(canvasRef.current.toDataURL());
             }
         }
-        push("/editvideo")
-    }
+        push("/editvideo");
+    };
 
     const onCancel = () => {
-        setIsModal(false)
-    }
+        setIsModal(false);
+    };
 
     const handleDragOver = (event: React.DragEvent<HTMLDivElement>) => {
         event.preventDefault();
@@ -60,16 +61,15 @@ export default function Upload() {
             if (inputRef.current.files.length > 0) {
                 if (handleFileSelect(inputRef.current.files[0]) === true) {
                     setIsModal(true);
-                    const blob = new Blob([files[0]], { type: 'image/png' })
+                    const blob = new Blob([files[0]], { type: "image/png" });
                     const url = URL.createObjectURL(blob);
                     setSrcVideoEdit(url);
-                }
-                else {
-                    alert('This is not video file.');
-                    inputRef.current.value = '';
+                } else {
+                    alert("This is not video file.");
+                    inputRef.current.value = "";
                 }
             } else {
-                console.log('Không có file được chọn');
+                console.log("Không có file được chọn");
             }
         }
     };
@@ -78,7 +78,7 @@ export default function Upload() {
         const files = event.target.files;
         if (files != null) {
             if (handleFileSelect(files[0]) === true) {
-                const blob = new Blob([files[0]], { type: 'image/png' })
+                const blob = new Blob([files[0]], { type: "image/png" });
                 const url = URL.createObjectURL(blob);
                 if (videoRef.current) {
                     videoRef.current.src = URL.createObjectURL(blob);
@@ -91,11 +91,11 @@ export default function Upload() {
                     };
                 }
                 setSrcVideoEdit(url);
-                setIsModal(true);
-            }
-            else {
-                alert('This is not video file.');
-                event.target.value = '';
+                setModalSelectCourse(!isModalSelectCourse);
+                // setIsModal(true);
+            } else {
+                alert("This is not video file.");
+                event.target.value = "";
             }
         }
     };
@@ -104,7 +104,7 @@ export default function Upload() {
         if (inputRef.current) {
             inputRef.current.click();
         }
-    }
+    };
 
     return (
         <>
@@ -116,7 +116,7 @@ export default function Upload() {
                     <div className="w-[100%] h-[90%] flex justify-center items-center">
                         <div
                             className={`w-[100%] h-[90%] border-dashed border-2 border-black flex justify-center items-center drag-drop-area 
-                            ${isDragging ? 'dragging' : ''}
+                            ${isDragging ? "dragging" : ""}
                             `}
                             onDragOver={handleDragOver}
                             onDragLeave={handleDragLeave}
@@ -124,23 +124,31 @@ export default function Upload() {
                         >
                             <div className="w-[50%] h-[90%]  flex flex-col justify-center items-center ">
                                 <div className="w-[100%] h-1/3 flex flex-col justify-center items-center">
-                                    <FaCloudUploadAlt className="h-full w-full flex items-center justify-center " size={100} />
+                                    <FaCloudUploadAlt
+                                        className="h-full w-full flex items-center justify-center "
+                                        size={100}
+                                    />
                                 </div>
                                 <div className="w-[100%] h-1/3 my-[10px] flex flex-col justify-center items-center">
                                     <div className="w-[100%] h-[60px] my-[10px]  flex justify-center items-end">
-                                        <p className="font-bold  text-2xl ">
-                                            Drag your file here
-                                        </p>
+                                        <p className="font-bold  text-2xl ">Drag your file here</p>
                                     </div>
                                     <div className="w-[100%] h-[60px] my-[10px]  flex justify-center items-start">
-                                        <p className="font-bold  text-2xl ">
-                                            Or
-                                        </p>
+                                        <p className="font-bold  text-2xl ">Or</p>
                                     </div>
                                 </div>
                                 <div className="w-[100%] h-1/3 my-[10px]  flex justify-center items-center">
-                                    <input className="hidden" type="file" accept="video/*" ref={inputRef} onChange={handleInputChange} />
-                                    <div onClick={handleButtonClick} className="w-[180px] h-[70px] border-2 border-solid boder-[#000000] flex justify-center items-center cursor-pointer">
+                                    <input
+                                        className="hidden"
+                                        type="file"
+                                        accept="video/*"
+                                        ref={inputRef}
+                                        onChange={handleInputChange}
+                                    />
+                                    <div
+                                        onClick={handleButtonClick}
+                                        className="w-[180px] h-[70px] border-2 border-solid boder-[#000000] flex justify-center items-center cursor-pointer"
+                                    >
                                         <p className="font-bold text-[25px]">Browser</p>
                                     </div>
                                 </div>
@@ -149,15 +157,10 @@ export default function Upload() {
                     </div>
                 </div>
             </div>
-            {isModal &&
-                <Modal
-                    title="Do you want to go to the video editing step?"
-                    onOk={onOk}
-                    onCancel={onCancel}
-                />
-            }
+            {isModal && <Modal title="Do you want to go to the video editing step?" onOk={onOk} onCancel={onCancel} />}
+            {isModalSelectCourse && <ModalSelectCourse isModalSelectCourse={isModalSelectCourse} setModalSelectCourse={setModalSelectCourse} setConfirmEditModal={setIsModal} />}
             <canvas ref={canvasRef} className="fixed z-[-10] opacity-0" />
             <video ref={videoRef} className="fixed z-[-10] opacity-0" />
         </>
-    )
+    );
 }
