+++ components/layout/SelectCourseLayout/index.tsx
@@ -5,6 +5,8 @@ import { Dispatch, ReactNode, SetStateAction } from "react";
 interface ISelectCourseLayoutProps {
     children: ReactNode;
     stepSelected: number;
+    setModalSelectCourse: React.Dispatch<React.SetStateAction<boolean>>;
+
     setStepSelected: Dispatch<SetStateAction<number>>;
     titleSteps: string[];
     setStepCreateCourse: Dispatch<SetStateAction<number>>;
@@ -15,12 +17,13 @@ export default function SelectCourseLayout({
     stepSelected,
     setStepSelected,
     titleSteps,
+    setModalSelectCourse,
     setStepCreateCourse,
 }: ISelectCourseLayoutProps) {
     return (
         <div className="w-[80%] h-[98%] bg-white px-4 pt-2 pb-6">
             <header className="h-[10%] ml-[3%] mr-[2%] flex items-center">
-                <SelectCourseHeader />
+                <SelectCourseHeader setModalSelectCourse={setModalSelectCourse} />
             </header>
             <div className="h-[90%]">
                 <div className="ml-[2%] h-[10%] flex items-center">
