--- pages/selectCourse/index.tsx
@@ -1,3 +1,0 @@
-export default function SelectCoursePage() {
-    return <div className="bg-[#7c7c7c] w-screen h-screen flex justify-center items-center"></div>;
-}
